+++ Assets/Game/Spaceship/Guns/Gun.cs
@@ -7,5 +7,5 @@ using VRC.Udon;
 public class Gun : UdonSharpBehaviour
 {
     public bool shooting = false;
-    public int damage = 2;
+    public float damage = 2f;
 }
